requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==11.0.0
numpy==2.0.2
torch==2.3.1
torchvision==0.18.1
